+++ components/MemoryGrid/index.js
