+++ lib/images.js
