+++ pages/index.js
@@ -1,8 +1,8 @@
-import MemoryGrid from "../src/components/MemoryGrid";
 import styled from "styled-components";
-import store from "../src/zustand/store";
-import InfoSection from "../src/components/InfoSection";
-import VictoryPopUp from "../src/components/VictoryPopUp";
+import store from "../zustand/store";
+import InfoSection from "../components/InfoSection";
+import VictoryPopUp from "../components/VictoryPopUp";
+import MemoryGrid from "../components/MemoryGrid";
 
 const StyledContainer = styled.div`
   display: flex;
